Move deps.ts and dev_deps.ts to root directory
